Add server mode conversion test
